Update in showtrace parser from parse of Integer to Long
git-svn-id: https://overture.svn.sourceforge.net/svnroot/overture/trunk@2338 2b2d4d09-c93c-414a-ad04-6e973f5c1cff

+++ core/showtrace/src/main/java/org/overturetool/traceviewer/parser/TraceParserVal.java
@@ -13,7 +13,7 @@ public class TraceParserVal
     {
     }
 
-    public TraceParserVal(int val)
+    public TraceParserVal(long val)
     {
         ival = val;
     }
@@ -33,7 +33,7 @@ public class TraceParserVal
         obj = val;
     }
 
-    public int ival;
+    public long ival;
     public double dval;
     public String sval;
     public Object obj;
